+++ generators/kubernetes-base.js
@@ -21,7 +21,7 @@ const _ = require('lodash');
 const shelljs = require('shelljs');
 const chalk = require('chalk');
 const crypto = require('crypto');
-const { defaultKubernetesConfig } = require('./generator-defaults');
+const { defaultKubernetesConfig } = require('./kubernetes/kubernetes-constants');
 const { loadFromYoRc } = require('./docker-base');
 const constants = require('./generator-constants');
 const { MICROSERVICE } = require('../jdl/jhipster/application-types');
