+++ generators/kubernetes/kubernetes-constants.js
@@ -1,0 +1,13 @@
+const kubernetesDefaultConfig = {
+  kubernetesServiceType: 'LoadBalancer',
+  monitoring: 'no',
+  istio: false,
+};
+
+const defaultKubernetesConfig = {
+  ...kubernetesDefaultConfig,
+};
+
+module.exports = {
+  defaultKubernetesConfig,
+};
